requirements.txt
blinker==1.9.0                # Used internally by Flask for signal support
click==8.2.1                  # CLI support (Flask dependency)
colorama==0.4.6               # For colored CLI output (Windows)
Flask==3.1.1                  # Main web framework
itsdangerous==2.2.0           # Secure cookie and URL handling (Flask core)
Jinja2==3.1.6                 # Flask templating engine
MarkupSafe==3.0.2             # Used by Jinja2
mysql-connector-python==9.3.0 # MySQL database driver
Werkzeug==3.1.3               # WSGI server and request/response handling
